Various Changes, see below:
Fixed bug where array references were being copied instead of cloned.
Renamed Executor class to ModelExecutor
Adjusted ModelExecutor to allow execution of a single block at a time
Changed the Virtual Machine to use the Model Executor instead of having duplicate code
Changed vector3 to floats instead of ints

+++ RobotInitial/Vector3.cs
@@ -5,11 +5,11 @@ using System.Text;
 
 namespace RobotInitial {
     struct Vector3 {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
 
-        public Vector3(int x, int y, int z) : this() {
+        public Vector3(float x, float y, float z) : this() {
             this.X = x;
             this.Y = y;
             this.Z = z;
